Code refactoring and organization
